+++ crates/factorio-rs/src/lib.rs
@@ -3,7 +3,7 @@
 pub use factorio_api::{self, event_type_to_name};
 pub use factorio_macros::{
     control, data, data_final_fixes, data_updates, event, export, item, locale, mod_settings,
-    recipe, settings, settings_final_fixes, settings_updates, shared,
+    recipe, settings, settings_final_fixes, settings_updates, shared, technology,
 };
 pub use factorio_macros::{
     control_mod, data_final_fixes_mod, data_mod, data_updates_mod, settings_final_fixes_mod,
@@ -26,7 +26,7 @@ pub mod prelude {
         control, control_mod, data, data_final_fixes, data_final_fixes_mod, data_mod, data_updates,
         data_updates_mod, event, export, item, locale, mod_settings, recipe, settings,
         settings_final_fixes, settings_final_fixes_mod, settings_mod, settings_updates,
-        settings_updates_mod, shared, shared_mod,
+        settings_updates_mod, shared, shared_mod, technology,
     };
     pub use factorio_api::prelude::*;